+++ src/types/index.ts
@@ -4,22 +4,18 @@ import type {
 	FastifyInstance,
 	RawServerBase,
 } from 'fastify';
-import type { JSONExportable } from '@piggly/ddd-toolkit';
+import type {
+	EnvironmentType,
+	JSONExportable,
+	LoggerService,
+} from '@piggly/ddd-toolkit';
 
 import type { Http2InsecureServer } from '@/www/fastify/Http2InsecureServer';
 import type { HttpInsecureServer } from '@/www/fastify/HttpInsecureServer';
 import type { Http2SecureServer } from '@/www/fastify/Http2SecureServer';
 import type { HttpSecureServer } from '@/www/fastify/HttpSecureServer';
-import type { LoggerService } from '@/services';
 
 /** Environment */
-export type EnvironmentType =
-	| 'homologation'
-	| 'development'
-	| 'production'
-	| 'sandbox'
-	| 'test';
-
 export type DefaultEnvironment = {
 	api: {
 		rest: {
@@ -47,11 +43,6 @@ export type FastifyServer<
 	| HttpSecureServer<AppEnvironment>;
 
 /** Fastify modifiers */
-export type FastifyModifierCallable<
-	Server extends RawServerBase,
-	AppEnvironment = DefaultEnvironment,
-> = (app: FastifyInstance<Server>, env: AppEnvironment) => Promise<void>;
-
 export interface FastifyAppliable<
 	Server extends RawServerBase,
 	AppEnvironment = DefaultEnvironment,
@@ -59,7 +50,21 @@ export interface FastifyAppliable<
 	apply: FastifyModifierCallable<Server, AppEnvironment>;
 }
 
+export type FastifyModifierCallable<
+	Server extends RawServerBase,
+	AppEnvironment = DefaultEnvironment,
+> = (app: FastifyInstance<Server>, env: AppEnvironment) => Promise<void>;
+
 /** Servers */
+
+export interface ApiServerInterface<
+	Server extends RawServerBase,
+	AppEnvironment extends DefaultEnvironment,
+> {
+	bootstrap: () => Promise<HttpServerInterface<Server, AppEnvironment>>;
+	getApp: () => FastifyInstance<Server>;
+	getEnv: () => AppEnvironment;
+}
 
 export type ApiServerOptions<
 	Server extends RawServerBase = RawServerDefault,
@@ -80,15 +85,6 @@ export type ApiServerOptions<
 	plugins: FastifyAppliable<Server, AppEnvironment>;
 	routes: FastifyAppliable<Server, AppEnvironment>;
 };
-
-export interface ApiServerInterface<
-	Server extends RawServerBase,
-	AppEnvironment extends DefaultEnvironment,
-> {
-	bootstrap: () => Promise<HttpServerInterface<Server, AppEnvironment>>;
-	getApp: () => FastifyInstance<Server>;
-	getEnv: () => AppEnvironment;
-}
 
 export interface HttpServerInterface<
 	Server extends RawServerBase,